refactor: modularize Chip
